+++ CodeStarterCamp_Week4/Activity.swift
@@ -1,0 +1,13 @@
+//
+//  Actibity.swift
+//  CodeStarterCamp_Week4
+//
+//  Created by 서진 on 2022/11/10.
+//
+
+import Foundation
+
+struct Activity {
+    let name: String
+    let action: (BodyCondition) -> Void
+}
